networkctl: Show public IP and other IP informations in azure env
In cloud enviroment instances are set public IP and private IP. Now
networkctl is not able to show the public IPs. On top of that
if more than one private ip is configured then except the IP which is
provided by DHCP others can't be fetched and configured via networkd.
This patch is adopting towards cloud env such as azure, gcp, elastic.
It's a minimum lib to pull the meta data and parse. The next set will
be allowing networkd to configure those private ips and subnets

https://docs.microsoft.com/en-us/azure/virtual-machines/linux/instance-metadata-service
https://github.com/SUSE-Enceladus/cloud-netconfig